Added Reproducibility and Checked it
Added by making sure to set random.seed() and SUMO's own --seed option.

Please maintain the Config import at the top of Anarcho.py, and maintain the random import at the top of Config.

I have not checked the results of not following those remarks. They may not be be useful.